refactor(zod-schemas): move zod-schemas to schemas folder and update imports
also rename type of schema values to QuickFormValues
