--- src/types/index.ts
@@ -1,6 +1,0 @@
-import { quickFormSchema, TQuickFormValues } from "./zod-schemas";
-
-export {
-    quickFormSchema,
-    type TQuickFormValues
-}
